requirements/dev.txt
black==22.12.0
mypy==0.991
mypy-extensions==0.4.3
pylint==2.15.9
PyYAML==6.0
pytest==7.2.0
types-beautifulsoup4==4.11.6.1
types-Markdown==3.4.2.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
